Extract Scala-JS-compatible zio-actors-api module
