+++ actors/src/main/scala/zio/actors/ActorRef.scala
@@ -1,49 +1,11 @@
 package zio.actors
 
-import java.io.{ IOException, ObjectInputStream, ObjectOutputStream, ObjectStreamException }
+import zio.actors.LocalActorSystem.resolvePath
 
+import java.io.{ IOException, ObjectInputStream, ObjectOutputStream, ObjectStreamException }
 import zio.nio.core.{ InetAddress, InetSocketAddress, SocketAddress }
 import zio.nio.core.channels.AsynchronousSocketChannel
 import zio.{ IO, Runtime, Task, UIO }
-
-/**
- * Reference to actor that might reside on local JVM instance or be available via remote communication
- *
- * @tparam F wrapper type constructing DSL
- */
-sealed trait ActorRef[-F[+_]] extends Serializable {
-
-  /**
-   * Send a message to an actor as `ask` interaction pattern -
-   * caller is blocked until the response is received
-   *
-   * @param fa message
-   * @tparam A return type
-   * @return effectful response
-   */
-  def ?[A](fa: F[A]): Task[A]
-
-  /**
-   * Send message to an actor as `fire-and-forget` -
-   * caller is blocked until message is enqueued in stub's mailbox
-   *
-   * @param fa message
-   * @return lifted unit
-   */
-  def !(fa: F[_]): Task[Unit]
-
-  /**
-   * Get referential absolute actor path
-   * @return
-   */
-  val path: UIO[String]
-
-  /**
-   * Stops actor and all its children
-   */
-  val stop: Task[List[_]]
-
-}
 
 /* INTERNAL API */
 
@@ -54,7 +16,6 @@ private[actors] object ActorRefSerial {
 private[actors] sealed abstract class ActorRefSerial[-F[+_]](private var actorPath: String)
     extends ActorRef[F]
     with Serializable {
-  import ActorSystemUtils._
 
   @throws[IOException]
   protected def writeObject1(out: ObjectOutputStream): Unit =
@@ -80,29 +41,6 @@ private[actors] sealed abstract class ActorRefSerial[-F[+_]](private var actorPa
   }
 
   override val path: UIO[String] = UIO(actorPath)
-}
-
-private[actors] final class ActorRefLocal[-F[+_]](
-  private val actorName: String,
-  actor: Actor[F]
-) extends ActorRefSerial[F](actorName) {
-  override def ?[A](fa: F[A]): Task[A] = actor ? fa
-
-  override def !(fa: F[_]): Task[Unit] = actor ! fa
-
-  override val stop: Task[List[_]] = actor.stop
-
-  @throws[IOException]
-  private def writeObject(out: ObjectOutputStream): Unit =
-    super.writeObject1(out)
-
-  @throws[IOException]
-  private def readObject(in: ObjectInputStream): Unit =
-    super.readObject1(in)
-
-  @throws[ObjectStreamException]
-  private def readResolve(): Object =
-    super.readResolve1()
 }
 
 private[actors] final class ActorRefRemote[-F[+_]](
