+++ actors/src/main/scala/zio/actors/ActorSystem.scala
@@ -1,16 +1,14 @@
 package zio.actors
 
-import java.io.{ ByteArrayInputStream, ByteArrayOutputStream, File, ObjectInputStream, ObjectOutputStream }
-import java.nio.ByteBuffer
-
-import zio.{ Chunk, IO, Promise, RIO, Ref, Task, UIO, ZIO }
-import zio.actors.Actor.{ AbstractStateful, Stateful }
+import zio.actors.LocalActorSystem.resolvePath
 import zio.actors.ActorSystemUtils._
-import zio.actors.ActorsConfig._
-import zio.clock.Clock
-import zio.nio.core.{ Buffer, InetAddress, SocketAddress }
+import zio.actors.config.{ Addr, Port, RemoteConfig }
 import zio.nio.core.channels.{ AsynchronousServerSocketChannel, AsynchronousSocketChannel }
+import zio.nio.core.{ Buffer, InetAddress, SocketAddress }
+import zio._
 
+import java.io._
+import java.nio.ByteBuffer
 import scala.io.Source
 
 /**
@@ -40,108 +38,16 @@ object ActorSystem {
 }
 
 /**
- * Context for actor used inside Stateful which provides self actor reference and actor creation/selection API
- */
-final class Context private[actors] (
-  private val path: String,
-  private val actorSystem: ActorSystem,
-  private val childrenRef: Ref[Set[ActorRef[Any]]]
-) {
-
-  /**
-   * Accessor for self actor reference
-   *
-   * @return actor reference in a task
-   */
-  def self[F[+_]]: Task[ActorRef[F]] = actorSystem.select(path)
-
-  /**
-   * Creates actor and registers it to dependent actor system
-   *
-   * @param actorName name of the actor
-   * @param sup       - supervision strategy
-   * @param init      - initial state
-   * @param stateful  - actor's behavior description
-   * @tparam S  - state type
-   * @tparam F1 - DSL type
-   * @return reference to the created actor in effect that can't fail
-   */
-  def make[R, S, F1[+_]](
-    actorName: String,
-    sup: Supervisor[R],
-    init: S,
-    stateful: Stateful[R, S, F1]
-  ): ZIO[R with Clock, Throwable, ActorRef[F1]] =
-    for {
-      actorRef <- actorSystem.make(actorName, sup, init, stateful)
-      children <- childrenRef.get
-      _        <- childrenRef.set(children + actorRef.asInstanceOf[ActorRef[Any]])
-    } yield actorRef
-
-  /**
-   * Looks up for actor on local actor system, and in case of its absence - delegates it to remote internal module.
-   * If remote configuration was not provided for ActorSystem (so the remoting is disabled) the search will
-   * fail with ActorNotFoundException.
-   * Otherwise it will always create remote actor stub internally and return ActorRef as if it was found.   *
-   *
-   * @param path - absolute path to the actor
-   * @tparam F1 - actor's DSL type
-   * @return task if actor reference. Selection process might fail with "Actor not found error"
-   */
-  def select[F1[+_]](path: String): Task[ActorRef[F1]] =
-    actorSystem.select(path)
-
-  /* INTERNAL API */
-
-  private[actors] def actorSystemName = actorSystem.actorSystemName
-
-  private[actors] def actorSystemConfig = actorSystem.config
-
-}
-
-/**
  * Type representing running instance of actor system provisioning actor herding,
  * remoting and actor creation and selection.
  */
 final class ActorSystem private[actors] (
-  private[actors] val actorSystemName: String,
-  private[actors] val config: Option[String],
-  private val remoteConfig: Option[RemoteConfig],
-  private val refActorMap: Ref[Map[String, Any]],
-  private val parentActor: Option[String]
-) {
-
-  /**
-   * Creates actor and registers it to dependent actor system
-   *
-   * @param actorName name of the actor
-   * @param sup       - supervision strategy
-   * @param init      - initial state
-   * @param stateful  - actor's behavior description
-   * @tparam S - state type
-   * @tparam F - DSL type
-   * @return reference to the created actor in effect that can't fail
-   */
-  def make[R, S, F[+_]](
-    actorName: String,
-    sup: Supervisor[R],
-    init: S,
-    stateful: AbstractStateful[R, S, F]
-  ): RIO[R with Clock, ActorRef[F]] =
-    for {
-      map          <- refActorMap.get
-      finalName    <- buildFinalName(parentActor.getOrElse(""), actorName)
-      _            <- if (map.contains(finalName)) IO.fail(new Exception(s"Actor $finalName already exists")) else IO.unit
-      path          = buildPath(actorSystemName, finalName, remoteConfig)
-      derivedSystem = new ActorSystem(actorSystemName, config, remoteConfig, refActorMap, Some(finalName))
-      childrenSet  <- Ref.make(Set.empty[ActorRef[Any]])
-      actor        <- stateful.makeActor(
-                        sup,
-                        new Context(path, derivedSystem, childrenSet),
-                        () => dropFromActorMap(path, childrenSet)
-                      )(init)
-      _            <- refActorMap.set(map + (finalName -> actor))
-    } yield new ActorRefLocal[F](path, actor)
+  actorSystemName: String,
+  config: Option[String],
+  remoteConfig: Option[RemoteConfig],
+  refActorMap: Ref[Map[String, Any]],
+  parentActor: Option[String]
+) extends LocalActorSystem(actorSystemName, config, remoteConfig, refActorMap, parentActor) {
 
   /**
    * Looks up for actor on local actor system, and in case of its absence - delegates it to remote internal module.
@@ -153,7 +59,7 @@ final class ActorSystem private[actors] (
    * @tparam F - actor's DSL type
    * @return task if actor reference. Selection process might fail with "Actor not found error"
    */
-  def select[F[+_]](path: String): Task[ActorRef[F]] =
+  override def select[F[+_]](path: String): Task[ActorRef[F]] =
     for {
       solvedPath                             <- resolvePath(path)
       (pathActSysName, addr, port, actorName) = solvedPath
@@ -161,48 +67,18 @@ final class ActorSystem private[actors] (
       actorMap <- refActorMap.get
 
       actorRef <- if (pathActSysName == actorSystemName)
-                    for {
-                      actorRef <- actorMap.get(actorName) match {
-                                    case Some(value) =>
-                                      for {
-                                        actor <- IO.effectTotal(value.asInstanceOf[Actor[F]])
-                                      } yield new ActorRefLocal(path, actor)
-                                    case None        =>
-                                      IO.fail(new Exception(s"No such actor $actorName in local ActorSystem."))
-                                  }
-                    } yield actorRef
+                    createActorLocalRef(path, actorName, actorMap)
                   else
                     for {
-                      address  <- InetAddress
-                                    .byName(addr.value)
-                                    .flatMap(iAddr => SocketAddress.inetSocketAddress(iAddr, port.value))
+                      address <- InetAddress
+                                   .byName(addr.value)
+                                   .flatMap(iAddr => SocketAddress.inetSocketAddress(iAddr, port.value))
                     } yield new ActorRefRemote[F](path, address)
     } yield actorRef
 
-  /**
-   * Stops all actors within this ActorSystem.
-   *
-   * @return all actors' unprocessed messages
-   */
-  def shutdown: Task[List[_]] =
-    for {
-      systemActors <- refActorMap.get
-      actorsDump   <- ZIO.foreach(systemActors.values.toList)(_.asInstanceOf[Actor[Any]].stop)
-    } yield actorsDump.flatten
-
-  /* INTERNAL API */
-
-  private[actors] def dropFromActorMap(path: String, childrenRef: Ref[Set[ActorRef[Any]]]): Task[Unit] =
-    for {
-      solvedPath          <- resolvePath(path)
-      (_, _, _, actorName) = solvedPath
-      _                   <- refActorMap.update(_ - actorName)
-      children            <- childrenRef.get
-      _                   <- ZIO.foreach_(children)(_.stop)
-      _                   <- childrenRef.set(Set.empty)
-    } yield ()
-
-  private def receiveLoop(address: ActorsConfig.Addr, port: ActorsConfig.Port): Task[Unit] =
+  override protected def newActorSystem[F[+_], S, R](finalName: String): LocalActorSystem =
+    new ActorSystem(actorSystemName, config, remoteConfig, refActorMap, Some(finalName))
+  private def receiveLoop(address: Addr, port: Port): Task[Unit]                          =
     for {
       addr      <- InetAddress.byName(address.value)
       address   <- SocketAddress.inetSocketAddress(addr, port.value)
@@ -258,38 +134,6 @@ final class ActorSystem private[actors] (
 /* INTERNAL API */
 
 private[actors] object ActorSystemUtils {
-
-  private val RegexName = "[\\w+|\\d+|(\\-_.*$+:@&=,!~';.)|\\/]+".r
-
-  private val RegexFullPath =
-    "^(?:zio:\\/\\/)(\\w+)[@](\\d+\\.\\d+\\.\\d+\\.\\d+)[:](\\d+)[/]([\\w+|\\d+|\\-_.*$+:@&=,!~';.|\\/]+)$".r
-
-  def resolvePath(path: String): Task[(String, Addr, Port, String)] =
-    RegexFullPath.findFirstMatchIn(path) match {
-      case Some(value) if value.groupCount == 4 =>
-        val actorSystemName = value.group(1)
-        val address         = Addr(value.group(2))
-        val port            = Port(value.group(3).toInt)
-        val actorName       = "/" + value.group(4)
-        IO.succeed((actorSystemName, address, port, actorName))
-      case None                                 =>
-        IO.fail(
-          new Exception(
-            "Invalid path provided. The pattern is zio://YOUR_ACTOR_SYSTEM_NAME@ADDRES:PORT/RELATIVE_ACTOR_PATH"
-          )
-        )
-    }
-
-  private[actors] def buildFinalName(parentActorName: String, actorName: String): Task[String] =
-    actorName match {
-      case ""            => IO.fail(new Exception("Actor actor must not be empty"))
-      case null          => IO.fail(new Exception("Actor actor must not be null"))
-      case RegexName(_*) => UIO.effectTotal(parentActorName + "/" + actorName)
-      case _             => IO.fail(new Exception(s"Invalid actor name provided $actorName. Valid symbols are -_.*$$+:@&=,!~';"))
-    }
-
-  def buildPath(actorSystemName: String, actorPath: String, remoteConfig: Option[RemoteConfig]): String =
-    s"zio://$actorSystemName@${remoteConfig.map(c => c.addr.value + ":" + c.port.value).getOrElse("0.0.0.0:0000")}$actorPath"
 
   def retrieveConfig(configFile: Option[File]): Task[Option[String]] =
     configFile.fold[Task[Option[String]]](Task.none) { file =>
