+++ api/src/main/scala/zio/actors/Supervisor.scala
@@ -1,9 +1,9 @@
 package zio.actors
 
 import zio.clock.Clock
-import zio.{ IO, RIO, Schedule, URIO, ZIO }
+import zio._
 
-private[actors] trait Supervisor[-R] {
+trait Supervisor[-R] {
   def supervise[R0 <: R, A](zio: RIO[R0, A], error: Throwable): ZIO[R0 with Clock, Unit, A]
 }
