+++ project/plugins.sbt
@@ -1,8 +1,12 @@
-addSbtPlugin("org.scalameta"      % "sbt-scalafmt"   % "2.4.0")
-addSbtPlugin("pl.project13.scala" % "sbt-jmh"        % "0.4.0")
-addSbtPlugin("com.eed3si9n"       % "sbt-buildinfo"  % "0.10.0")
-addSbtPlugin("org.scoverage"      % "sbt-scoverage"  % "1.6.1")
-addSbtPlugin("org.scalameta"      % "sbt-mdoc"       % "2.2.12")
-addSbtPlugin("ch.epfl.scala"      % "sbt-bloop"      % "1.4.5")
-addSbtPlugin("com.eed3si9n"       % "sbt-unidoc"     % "0.4.3")
-addSbtPlugin("com.geirsson"       % "sbt-ci-release" % "1.5.4")
+addSbtPlugin("org.scalameta"      % "sbt-scalafmt"             % "2.4.0")
+addSbtPlugin("pl.project13.scala" % "sbt-jmh"                  % "0.4.0")
+addSbtPlugin("com.eed3si9n"       % "sbt-buildinfo"            % "0.10.0")
+addSbtPlugin("org.scoverage"      % "sbt-scoverage"            % "1.6.1")
+addSbtPlugin("org.scalameta"      % "sbt-mdoc"                 % "2.2.12")
+addSbtPlugin("ch.epfl.scala"      % "sbt-bloop"                % "1.4.5")
+addSbtPlugin("com.eed3si9n"       % "sbt-unidoc"               % "0.4.3")
+addSbtPlugin("com.geirsson"       % "sbt-ci-release"           % "1.5.4")
+addSbtPlugin("org.scala-js"       % "sbt-scalajs"              % "1.3.1")
+addSbtPlugin("org.portable-scala" % "sbt-scalajs-crossproject" % "1.0.0")
+addSbtPlugin("org.scala-js"       % "sbt-jsdependencies"       % "1.0.0")
+addSbtPlugin("ch.epfl.scala"      % "sbt-scalajs-bundler"      % "0.20.0")
